rewritten the first part of the :: disambiguation chapter.
I also moved it up to just after the deferred kind-check chapter as it os on par with that as a tweak needed for UTPs to work well.

Then I added a short section on the  leading template and double typename template disambiguation cases.

Fixed the strange chapter levels at the end of the bikeshedding chapter.

Moved the jungle chapter far down and made it a level 1 chapter.